refactor: create AbstrctJdTest to make test easy
